Included dashboard template html
This template will be the basis for the page the user will see after logged in. We'll use this template for the dashboard, search results, reviews, etc etc. Always make sure the user is logged in before accessing the dasbhoard. Redirect to the log in page if user didn't log in.